Fix linting problems
