+++ package.json
@@ -9,7 +9,8 @@
     "react-dom": "^16.4.2",
     "react-router-dom": "^4.3.1",
     "react-scripts": "1.1.4",
-    "typeface-roboto": "0.0.54"
+    "typeface-roboto": "0.0.54",
+    "prop-types": "^15.6.2"
   },
   "devDependencies": {
     "eslint": "^5.5.0",
